extract shared structs
